refactor(CardList): ReservationCardList에서 CardList로 컴포넌트명 변경
